Extract jdk based implementation into its own module
