+++ settings.gradle
@@ -5,3 +5,4 @@ include 'var-serializer-guice'
 include 'var-test'
 include 'var-test-support'
 include 'var-undertow'
+include 'var-jdk'
