+++ var-core/build.gradle
@@ -1,9 +1,5 @@
 
 dependencies {
-    testImplementation group: 'junit', name: 'junit', version: '4.12'
-    implementation group: 'javax.inject', name: 'javax.inject', version: '1'
     api group: 'javax.servlet', name: 'javax.servlet-api', version: '4.0.1'
     implementation group: 'org.reflections', name: 'reflections', version: '0.9.11'
-    testImplementation group: 'org.mockito', name: 'mockito-all', version: '1.10.19'
-    implementation 'org.slf4j:slf4j-api:1.7.26'
 }
